requirements.txt
# mkdocs-material
#tensorflow
tensorflow-gpu
pandas
numpy
matplotlib
pillow
notebook
mlflow
pyyaml